ujwal room drama

+++ src/redux/ActionCreators.js
@@ -26,13 +26,17 @@ export const loginError = (message) => {
 
 // outpass functions
 export const postOutpass = (user, outpass) => async (dispatch) => {
-    console.log(user);
+    // console.log(user);
     outpass['uid'] = user.uid;
-    outpass['name']= user.name;
-    outpass['rollNum']= user.rollNum;
+    outpass['name'] = user.name;
+    outpass['rollNum'] = user.rollNum;
 
-    dispatch(requestOutpass());
     try {
+        dispatch(requestOutpass());
+
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
         await addDoc(collection(db, 'outpass'), outpass);
         dispatch(fetchOutpass(user));
     }
@@ -43,23 +47,24 @@ export const postOutpass = (user, outpass) => async (dispatch) => {
 
 export const fetchOutpass = (user) => async (dispatch) => {
 
-    dispatch(requestOutpass());
     try {
+        dispatch(requestOutpass());
+
         const querySnapshot = await getDocs(collection(db, "outpass"));
         let outpassArr = [];
 
         if (user === undefined)
-            throw Error("Error unauthorized");
+            throw Error("Error 401: Unauthorized");
         if (user.role === 'student') {
             const userid = user.uid;
             querySnapshot.forEach((doc) => {
                 if (doc.data().uid === userid) {
                     const _id = doc.id;
                     outpassArr.push({ _id, ...doc.data() });
-                }   
+                }
             })
         }
-        else if(user.role==='caretaker'){
+        else if (user.role === 'caretaker') {
             const hostel = user.hostelName;
             console.log(user);
             querySnapshot.forEach((doc) => {
@@ -67,7 +72,7 @@ export const fetchOutpass = (user) => async (dispatch) => {
                     outpassArr.push(doc.data());
             })
         }
-        else{
+        else {
             dispatch(outpassError('Error unauthorized'));
         }
         dispatch(receiveOutpass(outpassArr));
@@ -78,8 +83,14 @@ export const fetchOutpass = (user) => async (dispatch) => {
 }
 
 export const deleteOutpass = (outpass) => async (dispatch) => {
-    dispatch(requestOutpass());
     try {
+        dispatch(requestOutpass());
+
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
         const outpassRef = doc(db, "outpass", outpass.uid);
         await deleteDoc(outpassRef);
         dispatch(receiveOutpass(outpass));
@@ -88,12 +99,36 @@ export const deleteOutpass = (outpass) => async (dispatch) => {
     }
 }
 
+export const updateOutpass = (outpass) => async (dispatch) => {
+    try {
+        dispatch(requestOutpass());
+
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+        const outpassRef = firestore.doc(`outpass/${outpass._id}`)
+        await outpassRef.set(outpass, { merge: true });
+        const docOutpass = await getDoc(outpassRef);
+        dispatch(receiveOutpass(docOutpass.data()));
+
+    } catch (error) {
+        dispatch(outpassError(error.message));
+    }
+}
+
 // Bus functions
 export const postBus = (bus) => async (dispatch) => {
 
-    dispatch(requestBus());
 
     try {
+        dispatch(requestBus());
+
+
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
         await addDoc(collection(db, 'bus'), bus);
         dispatch(receiveBus(bus));
     }
@@ -104,13 +139,19 @@ export const postBus = (bus) => async (dispatch) => {
 
 export const fetchBus = () => async (dispatch) => {
 
-    dispatch(requestBus());
     try {
+
+        dispatch(requestBus());
+
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
         const querySnapshot = await getDocs(collection(db, "bus"));
         let busArr = [];
         querySnapshot.forEach((doc) => {
             const _id = doc.id;
-                    busArr.push({ _id, ...doc.data() });
+            busArr.push({ _id, ...doc.data() });
         })
         dispatch(receiveBus(busArr));
     }
@@ -121,8 +162,15 @@ export const fetchBus = () => async (dispatch) => {
 
 export const deleteBus = (bus) => async (dispatch) => {
     {
-        dispatch(requestBus());
         try {
+            dispatch(requestBus());
+
+            const user = auth.currentUser;
+            if (user === undefined)
+                throw Error("Error 401: Unauthorized");
+
+
+
             const busRef = doc(db, "bus", bus.busId);
             await deleteDoc(busRef);
             dispatch(receiveBus(bus));
@@ -133,10 +181,34 @@ export const deleteBus = (bus) => async (dispatch) => {
     };
 }
 
+export const updateBus = (bus) => async (dispatch) => {
+    try {
+        dispatch(requestBus());
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+        const busRef = firestore.doc(`bus/${bus._id}`)
+        await busRef.set(bus, { merge: true });
+        const docBus = await getDoc(busRef);
+        dispatch(receiveOutpass(docBus.data()));
+
+    } catch (error) {
+        dispatch(busError(error.message));
+    }
+}
+
 // Store functions
 export const postStore = (store) => async (dispatch) => {
-    dispatch(requestStore());
     try {
+
+        dispatch(requestStore());
+        const user = auth.currentUser;
+
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
+
         await addDoc(collection(db, 'store'), store);
         dispatch(receiveStore(store));
     }
@@ -147,8 +219,17 @@ export const postStore = (store) => async (dispatch) => {
 
 export const fetchStore = () => async (dispatch) => {
 
-    dispatch(requestStore());
+
     try {
+
+        dispatch(requestStore());
+
+        const user = auth.currentUser;
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
+
         const querySnapshot = await getDocs(collection(db, "store"));
         let storeArr = [];
         querySnapshot.forEach((doc) => {
@@ -169,8 +250,14 @@ export const postTicket = (ticket) => async (dispatch) => {
     const userid = user.uid;
     ticket['uid'] = userid;
 
-    dispatch(requestTicket());
     try {
+        dispatch(requestTicket());
+
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
+
         await addDoc(collection(db, 'ticket'), ticket);
         dispatch(receiveTicket(ticket));
     }
@@ -179,22 +266,27 @@ export const postTicket = (ticket) => async (dispatch) => {
     }
 };
 
-export const fetchTicket = () => async (dispatch) => {
+export const fetchTicket = (user) => async (dispatch) => {
 
-    dispatch(requestTicket());
     try {
+        dispatch(requestTicket());
+
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
         const querySnapshot = await getDocs(collection(db, "ticket"));
         let ticketArr = [];
-        const user = auth.currentUser;
+
         if (user !== undefined && user.role === 'student') {
             const userid = user.uid;
             querySnapshot.forEach((doc) => {
-                if (doc.data().uid === userid)
-                {const _id = doc.id;
-                ticketArr.push({ _id, ...doc.data() });}
+                if (doc.data().uid === userid) {
+                    const _id = doc.id;
+                    ticketArr.push({ _id, ...doc.data() });
+                }
             })
         }
-        else{
+        else {
             dispatch(ticketError('Error unauthorized'));
         }
         dispatch(receiveTicket(ticketArr));
@@ -205,63 +297,73 @@ export const fetchTicket = () => async (dispatch) => {
 }
 
 // Wallet functions
-export const postWallet = (wallet) => async (dispatch) => {
-    dispatch(requestWallet());
+export const postWallet = (user) => async (dispatch) => {
+
     try {
-        await addDoc(collection(db, 'wallet'), wallet);
-        dispatch(receiveWallet(wallet));
+
+        dispatch(requestWallet());
+
+        if (user === undefined)
+            throw Error("Error unauthorized");
+        const walletRef = firestore.doc(`wallet/${user.uid}`)
+        await walletRef.set({
+            tokenNo: 0,
+            uid: user.uid,
+        }, { merge: true }
+        )
+
+        const docUser = await getDoc(walletRef);
+
+        dispatch(receiveWallet(docUser.data()));
     }
     catch (error) {
         dispatch(walletError(error.message))
     }
 };
 
-export const fetchWallet = () => async (dispatch) => {
+export const fetchWallet = (user) => async (dispatch) => {
 
-    dispatch(requestWallet());
     try {
-        const user = auth.currentUser;
-        const userid = user.uid;
-        const querySnapshot = await getDocs(collection(db, "wallet"));
-        querySnapshot.forEach((doc) => {
-            if (doc.data().uid === userid)
-                dispatch(receiveWallet(doc.data()));
-        })
 
+        dispatch(requestWallet());
+
+
+        if (user === undefined)
+            throw Error("Error unauthorized");
+        const walletRef = firestore.doc(`wallet/${user.uid}`)
+        const docSnap = await getDoc(walletRef);
+
+        if (!docSnap.exists()) {
+            dispatch(postWallet(user));
+        }
+        else {
+            dispatch(receiveWallet(docSnap.data()));
+        }
     }
     catch (error) {
         dispatch(walletError(error.message))
     }
 }
 
-export const updateWallet = (wallet,tokens) => async (dispatch) => {
+export const updateWallet = (user, wallet, token) => async (dispatch) => {
 
-    dispatch(requestWallet());
-    try {    
-        const walletRef = doc(db,'wallet',wallet.uid);
+    try {
+        dispatch(requestWallet());
 
-        walletRef.get().then((doc) => {
-            if (doc.exists) {
-                console.log("Document data:", doc.data());
-            } else {
-                // doc.data() will be undefined in this case
-                console.log("No such document!");
-            }
-            console.log("KKK");
-        }).catch((error) => {
-            console.log("Error getting document:", error);
-        });
-    
-        // var bal = wallet.tokenBalance + tokens;
-        console.log(walletRef.uid+ "Atharba");
-        await updateDoc(walletRef, {
-            
-            // console.log(wallet.tokenBalance+'atharv');
-            tokenBalance:50
-        });
 
-        console.log(wallet);
-        dispatch(receiveWallet(wallet)); 
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+
+
+        const walletRef = firestore.doc(`wallet/${wallet.uid}`)
+
+        var newBal = wallet.tokenNo + token;
+        await walletRef.set({
+            tokenNo: newBal,
+        }, { merge: true }
+        )
+
+        dispatch(fetchWallet(user));
     }
     catch (error) {
         dispatch(walletError(error.message))
@@ -288,13 +390,19 @@ export const loginUser = (creds) => (dispatch) => {
                 querySnapshot.forEach((doc) => {
                     // doc.data() is never undefined for query doc snapshots
                     console.log(doc.id, " => ", doc.data());
+
                     dispatch(receiveLogin(doc.data()));
-                    dispatch(fetchOutpass(doc.data()));
+
+                    if (doc.data().role === "caretaker")
+                        dispatch(fetchOutpass(doc.data()));
+                    if (doc.data().role === "admin")
+                        dispatch(fetchSpecialBusRequest(doc.data()));
+
                 });
             })
-            .catch((error) => {
-                console.log("Error getting documents: ", error);
-            });
+                .catch((error) => {
+                    throw error;
+                });
         })
         .catch(error => dispatch(loginError(error.message)))
 };
@@ -319,7 +427,7 @@ export const logoutUser = () => (dispatch) => {
         // Sign-out successful.
     }).catch((error) => {
 
-        // An error happened.
+        throw Error(error);
     });
 
     localStorage.removeItem('user');
@@ -377,6 +485,9 @@ export const fetchUser = (user) => async (dispatch) => {
     const docUser = await getDoc(userRef);
     console.log(docUser.data());
 
+
+    dispatch(fetchWallet(docUser.data()));
+    dispatch(fetchTicket(docUser.data()));
     dispatch(receiveLogin(docUser.data()));
 }
 
@@ -394,15 +505,16 @@ export const googleLogin = () => (dispatch) => {
                 dispatch(loginError("Error 401: Unauthorized"));
             }
             else {
+
                 dispatch(fetchUser(user));
             }
 
-            if(user.role==='student'){
-                dispatch(fetchOutpass(user));
-            }
-            dispatch(fetchWallet(user));
-            const b = {uid : '1'};
-            dispatch(updateWallet(b, 8));
+            // if(user.role==='student'){
+            //     dispatch(fetchOutpass(user));
+            // }
+            // dispatch(fetchWallet(user));
+            // const b = {uid : '1'};
+            // dispatch(updateWallet(b, 8));
         })
         .catch((error) => {
             dispatch(loginError(error.message));
@@ -450,8 +562,8 @@ export const specialBusRequestError = (message) => {
 export const postSpecialBusRequest = (user, specialbusrequest) => async (dispatch) => {
     console.log(specialbusrequest);
     specialbusrequest['uid'] = user.uid;
-    specialbusrequest['name']= user.name;
-    specialbusrequest['email']= user.email;
+    specialbusrequest['name'] = user.name;
+    specialbusrequest['email'] = user.email;
 
 
     dispatch(requestSpecialBusRequest());
@@ -464,17 +576,23 @@ export const postSpecialBusRequest = (user, specialbusrequest) => async (dispatc
     }
 };
 
-export const fetchSpecialBusRequest = () => async (dispatch) => {
+export const fetchSpecialBusRequest = (user) => async (dispatch) => {
 
-    dispatch(requestSpecialBusRequest());
     try {
-        const querySnapshot = await getDocs(collection(db, "specialBusRequest"));
-        let specialBusArr = [];
-        querySnapshot.forEach((doc) => {
-            const _id = doc.id;
-            specialBusArr.push({ _id, ...doc.data() });
-        })
-        dispatch(receiveSpecialBusRequest(specialBusArr));
+        dispatch(requestSpecialBusRequest());
+
+        if (user !== undefined && user.role === 'admin') {
+            const querySnapshot = await getDocs(collection(db, "specialBusRequest"));
+            let specialBusArr = [];
+            querySnapshot.forEach((doc) => {
+                const _id = doc.id;
+                specialBusArr.push({ _id, ...doc.data() });
+            })
+            dispatch(receiveSpecialBusRequest(specialBusArr));
+        }
+        else {
+            throw Error("Error 401: Unauthorized");
+        }
     }
     catch (error) {
         dispatch(specialBusRequestError(error.message))
@@ -495,6 +613,20 @@ export const deleteSpecialBusRequest = (specialbusrequest) => async (dispatch) =
         }
 
     };
+}
+export const updateSpecialBus = (user, specialbus) => async (dispatch) => {
+    try {
+        dispatch(requestSpecialBusRequest());
+
+        if (user === undefined)
+            throw Error("Error 401: Unauthorized");
+        const specialBusRef = firestore.doc(`specialbus/${specialbus._id}`)
+        await specialBusRef.set(specialbus, { merge: true });
+        dispatch(fetchSpecialBusRequest());
+
+    } catch (error) {
+        dispatch(specialBusRequestError(error.message));
+    }
 }
 
 //outpass.js
@@ -597,9 +729,11 @@ export const walletError = (message) => {
 
 export const postSchedule = (schedule) => async (dispatch) => {
 
-    dispatch(requestSchedule());
-    dispatch(requestSpecialBusRequest());
+
     try {
+
+        dispatch(requestSchedule());
+        dispatch(requestSpecialBusRequest());
         await addDoc(collection(db, 'schedule'), schedule);
         dispatch(receiveSchedule(schedule));
     }
@@ -610,8 +744,11 @@ export const postSchedule = (schedule) => async (dispatch) => {
 
 export const fetchSchedule = () => async (dispatch) => {
 
-    dispatch(requestSchedule());
+
     try {
+
+        dispatch(requestSchedule());
+        dispatch(requestSpecialBusRequest());
         const querySnapshot = await getDocs(collection(db, "schedule"));
         let scheduleArr = [];
         querySnapshot.forEach((doc) => {
